restructure jobs
